Forward custom GA campaign params to Lever
This version is quite brainless, we just copy the URL's query string if
it's not empty. This is also saved into local storage, so that if
someone visits the page again w/o the query string it will be retained.

There are a lot of not-exactly edge cases where this doesn't do what we
would prefer, but as a simple first pass it should work.